Auto testing. yippee

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/WheelieThings/PathFollowerWrapper.java
@@ -20,8 +20,8 @@ public class PathFollowerWrapper {
     private final PID hPID;
     private final ElapsedTime pidTimer;
 
-    private static final double mP = (double) 1/24, mI = 0, mD = 0,
-            hP = 1./Math.PI, hI = 0, hD = 0,
+    private static final double mP = (double) 1 / 24, mI = 0, mD = 0,
+            hP = (double) 1 / Math.PI, hI = 0, hD = 0,
             mMaxI = 0.25, hMaxI = 0.1;
 
     //The max speed of the motors
